use low_memory option to avoid DType warning
It's not efficient but will do for now.
Sometimes a CSV file can contain string and numbers and it's difficult
for pandas to guess the type of each columns by looking at the first
rows (a row can contain -1 and then string data, or even being empty).

So we use the low_memory=False option which will have to read the
all file to guess the type of each columns.
Let see if we find a better way in the future.